Implement live statistics
Configurable via the config file of course,
fetches movements, controllers and ATIS from the datafile
and displays this in a live-updating set of embeds.

I chose to use three embeds, shown when there is data,
to avoid ever hitting the fields-per-embed limit.

Theoretically still possible to hit it if there's more thatn 25 ATIS'es
or controllers, but this seems unlikely.

Update runs on startup and every 15 minutes afterwards.