Add initial Scenario command sketch
I attempted to put a lot of the input in a modal,
but for now this appears undoable due to the limit on 5 text fields.

Custom intensity will _not_ be supported in the immediate future,
again because Discord limits us to 5 free-text inputs after the command.

I will probably end up having to pull apart this ts into several files,
just to keep it doable.

Next steps are per-airport runway config selection

Of course at some point we'll have to decide whether we'll be porting
the actual generation script over or not