Fix not all roles being added to event threads
This issue boils down to not having access to all members of a role.
The only way around this that I found,
is allowing the GUILD_MEMBERS privileged intent (with all due
consequences) and explicitly fetching all guild members before checking
for role membership.

This slows down the command somewhat and adds a privileged intent,
but seems necessary.